requirements.txt

pillow
pandas
openpyxl